some small cleanups, refactored email function to be accessible to other modules
